Fix issue with status icon resizing, move wait time to a component
